Periodic merge with OT

See merge request cms_tk_ph2/Ph2_ACF!594

+++ DQMUtils/RD53SCurveHistograms.h
@@ -34,7 +34,7 @@ class SCurveHistograms : public DQMHistogramBase
     bool fill(std::string& inputStream) override;
     void reset() override{};
 
-    void fillOccupancy(const DetectorDataContainer& OccupancyContainer, int DELTA_VCAL);
+    void fillOccupancy(const DetectorDataContainer& OccupancyContainer, uint16_t DELTA_VCAL);
     void fillThrAndNoise(const DetectorDataContainer& ThrAndNoiseContainer);
 
     bool AreHistoBooked = false;
